+++ scout/commands/view/case.py
@@ -27,6 +27,7 @@ def cases(institute, display_name, case_id, nr_variants, variants_treshold, simi
     """Display cases from the database"""
     LOG.info("Running scout view institutes")
     adapter = store
+    name_query = None
 
     models = []
     if case_id:
@@ -34,7 +35,11 @@ def cases(institute, display_name, case_id, nr_variants, variants_treshold, simi
         if case_obj:
             models.append(case_obj)
         if similar:
-            similar = adapter.get_similar_cases(case_obj)
+            hpo_terms = []
+            for term in case_obj.get("phenotype_terms", []):
+                hpo_terms.append(term.get("phenotype_id"))
+
+            similar = adapter.cases_by_phenotype(hpo_terms, case_obj["owner"], case_obj["_id"])
             if not similar:
                 LOG.info("No more cases with phenotypes found")
                 return
@@ -44,7 +49,9 @@ def cases(institute, display_name, case_id, nr_variants, variants_treshold, simi
             return
 
     else:
-        models = adapter.cases(collaborator=institute, name_query=display_name)
+        if display_name:
+            name_query = f"case:{display_name}"
+        models = adapter.cases(collaborator=institute, name_query=name_query)
         models = [case_obj for case_obj in models]
 
     if not models: