+++ tests/adapter/mongo/test_case_handling.py
@@ -171,13 +171,13 @@ def test_get_cases_no_HPO(adapter, case_obj):
     adapter.case_collection.insert_one(case_obj)
 
     # WHEN providing an empty value for term HP:
-    name_query = "HP:"
+    name_query = "exact_pheno:"
     # Then case should be returned
     cases = adapter.cases(collaborator=case_obj["owner"], name_query=name_query)
     assert sum(1 for i in cases) == 1
 
     # WHEN providing an empty value for phenotype group:
-    name_query = "PG:"
+    name_query = "pheno_group:"
     # Then case should be returned
     cases = adapter.cases(collaborator=case_obj["owner"], name_query=name_query)
     assert sum(1 for i in cases) == 1
@@ -193,13 +193,13 @@ def test_get_cases_no_HPO(adapter, case_obj):
         },
     )
     # WHEN providing an empty value for term HP:
-    name_query = "HP:"
+    name_query = "exact_pheno:"
     # Then case should NOT be returned
     cases = adapter.cases(collaborator=case_obj["owner"], name_query=name_query)
     assert sum(1 for i in cases) == 0
 
     # WHEN providing an empty value for phenotype group:
-    name_query = "PG:"
+    name_query = "pheno_group:"
     # Then case should NOT be returned
     cases = adapter.cases(collaborator=case_obj["owner"], name_query=name_query)
     assert sum(1 for i in cases) == 0
@@ -211,7 +211,7 @@ def test_get_cases_no_assignees(real_adapter, case_obj):
     assert adapter.case_collection.find_one() is None
     adapter.case_collection.insert_one(case_obj)
     # WHEN retreiving an existing case from the database
-    result = adapter.cases(name_query="john")
+    result = adapter.cases(name_query="user:john")
     # THEN we should get the correct case
     assert sum(1 for i in result) == 0
 
@@ -228,7 +228,7 @@ def test_get_cases_display_name(real_adapter, case_obj):
     adapter.case_collection.insert_one(other_case)
 
     # WHEN retreiving cases by partial display name
-    result = adapter.cases(name_query="643")
+    result = adapter.cases(name_query="case:643")
     # THEN we should get the correct case
     assert sum(1 for i in result) == 1
 
@@ -239,7 +239,7 @@ def test_get_cases_existing_individual(real_adapter, case_obj):
     assert adapter.case_collection.find_one() is None
     adapter.case_collection.insert_one(case_obj)
     # WHEN retreiving cases by partial individual name
-    result = adapter.cases(name_query="NA1288")
+    result = adapter.cases(name_query="case:NA1288")
     # THEN we should get the correct case
     assert sum(1 for i in result) == 1
 
@@ -256,7 +256,7 @@ def test_get_cases_assignees(real_adapter, case_obj, user_obj):
     adapter.case_collection.insert_one(case_obj)
 
     # WHEN retreiving cases by partial individual name
-    result = adapter.cases(name_query="john")
+    result = adapter.cases(name_query="user:john")
     # THEN we should get the correct case
     assert sum(1 for i in result) == 1
 
@@ -273,7 +273,7 @@ def test_get_cases_non_existing_assignee(real_adapter, case_obj, user_obj):
     adapter.case_collection.insert_one(case_obj)
 
     # WHEN retreiving cases by partial individual name
-    result = adapter.cases(name_query="damien")
+    result = adapter.cases(name_query="user:damien")
     # THEN we should get the correct case
     assert sum(1 for i in result) == 0
 
@@ -321,13 +321,13 @@ def test_get_cases_empty_causatives(adapter, case_obj):
     assert sum(1 for i in result) == 0
 
 
-def test_get_cases_non_existing_individual(real_adapter, case_obj):
+def test_get_cases_non_existing_display_name(real_adapter, case_obj):
     adapter = real_adapter
     # GIVEN an empty database (no cases)
     assert adapter.case_collection.find_one() is None
     adapter.case_collection.insert_one(case_obj)
     # WHEN retreiving cases by partial display name
-    result = adapter.cases(name_query="hello")
+    result = adapter.cases(name_query="case:hello")
     # THEN we should get the correct case
     assert sum(1 for i in result) == 0
 
@@ -522,7 +522,7 @@ def test_update_case_rerun_status(adapter, case_obj, institute_obj, user_obj):
     assert res["status"] == "inactive"
 
 
-def test_get_similar_cases(hpo_database, test_hpo_terms, case_obj):
+def test_cases_by_phenotype(hpo_database, test_hpo_terms, case_obj):
     adapter = hpo_database
 
     # Make sure database contains HPO terms
@@ -559,7 +559,8 @@ def test_get_similar_cases(hpo_database, test_hpo_terms, case_obj):
     adapter.case_collection.insert_one(case_3)
     assert sum(1 for i in adapter.case_collection.find()) == 3
 
-    similar_cases = adapter.get_similar_cases(case_obj)
+    hpo_query_terms = [term["phenotype_id"] for term in test_hpo_terms]
+    similar_cases = adapter.cases_by_phenotype(hpo_query_terms, case_obj["owner"], case_obj["_id"])
     # make sure that the function returns a list of tuples
     assert isinstance(similar_cases, list)
     assert isinstance(similar_cases[0], tuple)
@@ -597,8 +598,7 @@ def test_get_similar_cases_by_name_query(hpo_database, test_hpo_terms, case_obj)
     assert sum(1 for i in adapter.case_collection.find()) == 2
 
     # WHEN querying for a similar case
-    name_query = "similar:{}".format(case_obj["display_name"])
-
+    name_query = f"similar_case:{case_obj['display_name']}"
     # THEN one case should be returned
     cases = list(adapter.cases(collaborator=case_obj["owner"], name_query=name_query))
     assert len(cases) == 1