+++ CHANGELOG.md
@@ -7,6 +7,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 ## [x.x.x]
 ### Added
 - Display number of filtered variants vs number of total variants in variants page
+- Search case by HPO terms
 
 ### Fixed
 - Bug occurring when rerun is requested twice
@@ -16,6 +17,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
 - Updated the documentation on how to create a new software release
 - Genome build-aware cytobands coordinates
 - Styling update of the Matchmaker card
+- Select search type in case search form
 
 
 ## [4.19]